requirements.txt
requests==2.31.0
python-dotenv==1.0.0
ccxt==4.2.15
beautifulsoup4==4.12.3
torch==2.2.0
transformers==4.37.2
accelerate==0.27.2 